requirements.txt
Pillow
arabic-reshaper
python-bidi
pyyaml
